Fix buffer size mismatch in QSSGRhiQuadRenderer::ensureBuffers()
... the wrong size will later backfire in staticUpload() as it will
use the size of QSSGRhiBuffer when copying from the stack buffer.

+++ src/runtimerender/qssgrhiquadrenderer.cpp
@@ -71,7 +71,7 @@ void QSSGRhiQuadRenderer::ensureBuffers(QSSGRhiContext *rhiCtx, QRhiResourceUpda
                                    QRhiBuffer::Immutable,
                                    QRhiBuffer::IndexBuffer,
                                    0,
-                                   6 * sizeof(float),
+                                   6 * sizeof(quint16),
                                    QRhiCommandBuffer::IndexUInt16);
         m_ibuf->buffer()->setName(QByteArrayLiteral("quad index buffer"));
         const quint16 buf[] = { 0, 1, 2, 0, 2, 3 };
